examples/python/v2/servers/mainnet/pyproject.toml
[project]
name = "x402-mainnet-example-v2"
version = "0.1.0"
description = "Example of using a fastapi server with x402 middleware"
requires-python = ">=3.10"
dependencies = [
    "x402",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.0",
    "cdp-sdk>=1.35.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
package = false

[tool.uv.sources]
x402 = { path = "../../../../../python/x402", editable = true }
